add role for fixing directory permissions
Tahoe has a problem with conflicts between roles. `gcsfuse` creates
`/var/www/letsencrypt`, which implicitly creates `/var/www` if it
doesn't already exist. `nginx` creates `/var/www` but because we do
other things in that role, in Tahoe's case it needs to run after
gcsfuse. That leads to incorrect permissions/ownership on
`/var/www`. We can't fix the ownership in the gcsfuse role because it
runs potentially before nginx has created the user that it needs. The
cleanest solution I could come up with (without completely refactoring
our entire deployment) is to add a very simple role that we can run
after both of them to fix up the permissions.