force permissions on nginx webroot parent directory
Depending on the order that roles get run during an initial deployment
when nginx, gcsfuse, and letsencrypt are all involved, we can end up
with permissions on `/var/www` that prevent nginx from accessing the
directories within. This just basically adds an extra `chown` to
prevent that. In cases where the bug doesn't happen, it should be
harmless.