scriabin: write metrics to node_exporter text collector dir
now that we have node_exporter installed (at least everywhere that
scriabin is installed), we ought to be able to rely on it for
exporting metrics instead of going through the pushgateway.

This adds a little bit of code to do that.

For now, we leave in the pushgateway parts so it's actually exporting
metrics both ways. This gives us some overlap so we can let the new
metrics (collected via node_exporter) populate, then update our
Grafana graphs to use them. Then we can remove the old pushgateway
code once those metrics are no longer being used.