remove nonempty option from gcsfuse mount
no longer supported in current versions of gcsfuse. See: https://github.com/GoogleCloudPlatform/gcsfuse/issues/424

Running into this on Juniper. We'll also hit it if we have to deploy
new servers for Tahoe hawthorn since we don't pin the version and a
freshly provisioned server would get the newer gcsfuse and break.

`-o nonempty` just tells it that it's OK to mount a directory even if
data already exists in that directory. That *shouldn't* occur anyway
in our regular case. If we find that we're trying to mount a non-empty
directory, that probably indicates a bug elsewhere in our setup that
should be investigated and fixed rather than papered over with this.